Add --dev flag for ssh cloning (#7)

+++ cave.sh
@@ -191,7 +191,13 @@ create_cave() { # Create a cave app instance in folder $1
     printf "Cannot create app '$1': This folder already exists in the current directory\n"
     exit 1
   fi
-  local CLONE_URL="${HTTPS_URL}"
+  local DEV_IDX=$(indexof --dev "$@")
+  if [ ! "${DEV_IDX}" = "-1" ]; then
+    local CLONE_URL="${SSH_URL}"
+  else
+    local CLONE_URL="${HTTPS_URL}"
+  fi
+
   local VERSION_IDX=$(indexof --version "$@")
   local offset=$(echo "${VERSION_IDX} + 2" | bc -l)
 
@@ -212,7 +218,9 @@ create_cave() { # Create a cave app instance in folder $1
     $PYTHON3_BIN -m pip install virtualenv
   fi
   cd "$1"
-  git remote rm origin
+  if [ "${DEV_IDX}" = "-1" ]; then
+    git remote rm origin
+  fi
   $PYTHON3_BIN -m virtualenv venv
 
   # Activate venv and install requirements
